requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.48
httpx==0.27.0
pyyaml==6.0.1
python-multipart==0.0.9
aiofiles==23.2.1
pycryptodome==3.20.0
Pillow==11.2.1
